+++ src/Component/OpenApi3/Tests/fixtures/parameters/expected/Endpoint/TestFormFileParameters.php
@@ -27,6 +27,9 @@ class TestFormFileParameters extends \Jane\Component\OpenApi3\Tests\Expected\Run
             $formParameters = $serializer->normalize($this->body, 'json');
             foreach ($formParameters as $key => $value) {
                 $value = is_int($value) ? (string) $value : $value;
+                if (is_array($value)) {
+                    $value = $serializer->serialize($value, 'json');
+                }
                 $bodyBuilder->addResource($key, $value);
             }
             return [['Content-Type' => ['multipart/form-data; boundary="' . ($bodyBuilder->getBoundary() . '"')]], $bodyBuilder->build()];